Re-factored resample into it's own module for easier maintenance
